Completed admin visit functionality and implemented visit log for doctor and patient in admin views

+++ database/migrations/2019_11_12_174914_create_visits_table.php
@@ -16,7 +16,7 @@ class CreateVisitsTable extends Migration
         Schema::create('visits', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('notes', 255);
-            $table->float('duration');
+            $table->integer('duration');
             $table->timestamps();
             
             $table->bigInteger('doctor_id')->unsigned();
